requirements.txt
# Core
pyrtlsdr>=0.3.0
pyrtlsdrlib>=0.0.2
pyhackrf2>=1.0.0
numpy>=1.23
scipy>=1.9

# API server
fastapi>=0.111
uvicorn>=0.30
pydantic>=2.0
